Move `FolderClass` to `:legacy:core`
We only keep `FolderClass` around for settings migrations.
