+++ legacy/core/src/main/java/com/fsck/k9/preferences/FolderSettingsDescriptions.java
@@ -7,7 +7,7 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.TreeMap;
 
-import com.fsck.k9.mail.FolderClass;
+import com.fsck.k9.preferences.legacy.FolderClass;
 import com.fsck.k9.preferences.Settings.BooleanSetting;
 import com.fsck.k9.preferences.Settings.EnumSetting;
 import com.fsck.k9.preferences.Settings.SettingsDescription;
