+++ legacy/core/src/main/java/com/fsck/k9/preferences/legacy/FolderClass.kt
@@ -1,0 +1,8 @@
+package com.fsck.k9.preferences.legacy
+
+internal enum class FolderClass {
+    NO_CLASS,
+    INHERITED,
+    FIRST_CLASS,
+    SECOND_CLASS,
+}
