--- mail/common/src/main/java/com/fsck/k9/mail/FolderClass.java
@@ -1,5 +1,0 @@
-package com.fsck.k9.mail;
-
-public enum FolderClass {
-    NO_CLASS, INHERITED, FIRST_CLASS, SECOND_CLASS
-}
